% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_timeNumber.R
\name{add_timeNumber}
\alias{add_timeNumber}
\title{add_timeNumber}
\usage{
add_timeNumber(data, tp_col_nm = "timestamp")
}
\arguments{
\item{data}{data with a time point column.}

\item{tp_col_nm}{column name of the data column containing the time point information.}
}
\value{
data with the extra timeNumber column added
}
\description{
Add a time number (1, 2, ...) column to data including time points.
}
