% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convert_old_unit_row_col.R
\name{convert_old_unit_row_col}
\alias{convert_old_unit_row_col}
\title{convert_old_unit_row_col}
\usage{
convert_old_unit_row_col(x)
}
\arguments{
\item{x}{vector of 'old unit' LeasyScan plot positions}
}
\value{
data.frame with row and column position
}
\description{
Convert the old unit information (A-1-1) into row column info
A-1 (first column), ...-...- 1 (first row or sector).
}
