\name{LS_new_unit_row_lk}
\alias{LS_new_unit_row_lk}
\docType{data}
\title{
  Look up for LeasyScan data conversion between new unit (XXX:YY:Z) and row number
}
\description{
  Example plant eye data.
}
\usage{data("LS_new_unit_row_lk")}
\format{
  A data frame with ...
}

\examples{
  data(LS_new_unit_row_lk)
}
\keyword{datasets}
