\name{LS_unit_col_lk}
\alias{LS_unit_col_lk}
\docType{data}
\title{
  Look up for LeasyScan data conversion between (old) unit (A-XX:YY:Z) and column number
}
\description{
  Example plant eye data.
}
\usage{data("LS_unit_col_lk")}
\format{
  A data frame with ...
}

\examples{
  data(LS_unit_col_lk)
}
\keyword{datasets}
