% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mdf_raw_pe_colnames.R
\name{mdf_raw_pe_colnames}
\alias{mdf_raw_pe_colnames}
\title{mdf_raw_pe_colnames}
\usage{
mdf_raw_pe_colnames(colnames)
}
\arguments{
\item{colnames}{String vector of column names.}
}
\value{
Modified column names.
}
\description{
A function to detect the desired column names: digital biomass,
height, leaf angle, etc. and format them with standard names accepted by the
\cite{\link{median_data_computation function}}.
}
