% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_median_computation.R
\name{median_computation}
\alias{median_computation}
\title{median_computation}
\usage{
median_computation(data)
}
\arguments{
\item{data}{data.frame with a "timePoint" column describing the different
timestep of the data, the experimental design information like replicate,
block, row, columns, etc. This part finish with the "genotype" column indicating
the genotype treatment of the experiment. Then all remaining column are
numeric trait values.}
}
\value{
data.frame with all design element columns reduced and all the traits
with median values per day
}
\description{
function to calculate the median data of each plot (sector)
for a certain number of traits
}
