% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_outlier_boxplot_detect.R
\name{outlier_boxplot_detect}
\alias{outlier_boxplot_detect}
\title{outlier_boxplot_detect}
\usage{
outlier_boxplot_detect(data)
}
\arguments{
\item{data}{data.frame with a "timePoint" column describing the different
timestep of the data, the experimental design information like replicate,
block, row, columns, etc. This part finish with the "genotype" column indicating
the genotype treatment of the experiment. Then all remaining column are
numeric trait values.}
}
\value{
data with outliers removed
}
\description{
Detececion of the outliers using boxplot method.
All data above 1.5 third quartile and below 1.5 first quartile
are considered as outliers and set as missing.
}
