% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_growth_curve.R
\name{plot_growth_curve}
\alias{plot_growth_curve}
\title{plot_growth_curve}
\usage{
plot_growth_curve(
  raw,
  linear,
  logistic,
  geno_sel = NULL,
  main = "Growth curve plot"
)
}
\arguments{
\item{raw}{matrix raw observed data obtained with function growth_curve_fit}

\item{linear}{matrix linear trend fitted data obtained with function growth_curve_fit}

\item{logistic}{matrix logistic trend fitted data obtained with function growth_curve_fit}

\item{geno_sel}{numeric vector selected genotypes}

\item{main}{plot title}
}
\value{
growth curve plot
}
\description{
Plot growth curve fundtions for selected genotype.
}
