% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_h2_trend.R
\name{plot_h2_trend}
\alias{plot_h2_trend}
\title{plot_h2_trend}
\usage{
plot_h2_trend(h2_LC, h2_PE, tp_conv = "day_gap", main = "heritability trend")
}
\arguments{
\item{h2_LC}{heritability results for loadcell (transpiration) data obtained
from function spatial_adjustment.}

\item{h2_PE}{heritability results for loadcell (transpiration) data obtained
from function spatial_adjustment.}

\item{tp_conv}{specify the way tp are transformed in x axis. Either, follow
the number of days in the date "day_gap" or tp is converted into 1, 2, ...
"consecutive"}

\item{main}{graph title}
}
\value{
plot of the heritabilities per trait over time.
}
\description{
Plot heritability trend
}
