% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_QTL_sup.R
\name{plot_QTL_sup}
\alias{plot_QTL_sup}
\title{plot_QTL_sup}
\usage{
plot_QTL_sup(data, traits, threshold = 3)
}
\arguments{
\item{data}{QTL profile data obtained with QTL_initial_scan function}

\item{traits}{character vector of selected traits.}

\item{threshold}{QTL selection threshold}
}
\value{
The return value, if any, from executing the function.
}
\description{
Superimposed plot of QTL profiles
}
