% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{plot_trend}
\usage{
plot_trend(data, trait = NULL, main = NULL, genotype = FALSE)
}
\arguments{
\item{data}{data.frame with a "timeNb" column describing the different
timestep of the data (1, 2, ...), the trait and plotId column. Object
obtained with the function \code{\link{median_computation}}}

\item{trait}{character string specifiying the trait.}

\item{main}{Title of the graph.}

\item{genotype}{logical value specifying if the grouping factor is genotype.
By default values are grouped by plot.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Plot the trend of a trait value.
}
