% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_QTL_initial_scan.R
\name{QTL_initial_scan}
\alias{QTL_initial_scan}
\title{QTL_initial_scan}
\usage{
QTL_initial_scan(
  mppData,
  traits,
  Q.eff = "par",
  threshold = 3,
  window = 50,
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{mppData}{mppData object containing phenotypic and genotypic data.}

\item{traits}{character vector list of traits.}

\item{Q.eff}{type of QTL effect}

\item{threshold}{numeric threshold value for QTL selection}

\item{window}{numeric value for exclusion window in the QTL selection procedure.}

\item{n.cores}{number of cores.}

\item{verbose}{logical value indicating if traits indicator should be printed}
}
\value{
The return value, if any, from executing the function.
}
\description{
Perform an initial QTL scan on a list of traits.
}
