% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatial_adjustment.R
\name{spatial_adjustment}
\alias{spatial_adjustment}
\title{spatial_adjustment}
\usage{
spatial_adjustment(
  TP,
  traits,
  timePoints = NULL,
  extraFixedFactors = NULL,
  geno.decomp = NULL,
  what = c("random", "fixed"),
  h2_comp = TRUE,
  useCheck = FALSE,
  useRepId = FALSE,
  engine = c("SpATS", "asreml"),
  spatial = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{TP}{object of class TP. see documentation from statgenHTP package.}

\item{traits}{vector of traits. must have the same identifiers as in the TP
object.}

\item{timePoints}{numeric vector of time points for which the adjustment
should be performed.}

\item{extraFixedFactors}{character vector specifying the term of the design
that should be included in the adjustment.}

\item{geno.decomp}{see fitModels function from the statGenHTP package.}

\item{what}{character string specifying if the genotype term should be fitted
as random or as fixed.}

\item{h2_comp}{logical value specifying if heritability should be computed.
If the geno term (what) is fixed. It will change to "random".}

\item{useCheck}{logical value specifying if check term should be included
in the model. For more detail see fitModels function from the statGenHTP package.}

\item{useRepId}{logical value. see fitModels function from the statGenHTP package.}

\item{engine}{specify if SpATS or asreml package should be used for the adjustment.}

\item{spatial}{see fitModels function from the statGenHTP package.}

\item{quiet}{logical value specifying if the timepoint should be printed}
}
\value{
list with three data frame containing the: a) adjusted plot data;
b) the adjusted genotype data; c) the way the adjustment was performed;
d) table with heritability values (if spatial adjustment could be calculated).
}
\description{
Function performing spatial adjustment for a list of traits data
stored in a TP object. The adjustment is calculated for each trait x timepoint.
The function try to fit model from the most complex to the most simple:
a) (if fixed effect specified) fixed effect + spatial correction; b)
(if previous fail) spatial correction only; c) (if everything fail)
simple mean over genotype or row values for plot level. The simple average
is not performed if the genotype term is specified as random because the
difference of range between term averaged and term adjusted from a random term
can be too large. The single model computation is done using the util function
spatial_adjustment_ij.
}
