% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_timepoint_perc_missing.R
\name{timepoint_prop_non_missing}
\alias{timepoint_prop_non_missing}
\title{timepoint_prop_non_missing}
\usage{
timepoint_prop_non_missing(data)
}
\arguments{
\item{data}{data.frame with a "timePoint" column describing the different
timestep of the data, the experimental design information like replicate,
block, row, columns, etc. This part finish with the "genotype" column indicating
the genotype treatment of the experiment. Then all remaining column are
numeric trait values.}
}
\value{
vector of proportion of non missing sector per day.
}
\description{
Determine the proportion of plot with at least one observation
# for the set of trait over the time points of the time series.
}
